src/data_fetch_from_api/polygonapi (tech stocks)/requirements.txt
google-cloud-storage
google-auth
pandas
google-cloud
datetime
